+++ test/Main.hs
@@ -1,6 +1,7 @@
 module Main (main) where
 
 import HoYo.Test.Bookmark
+import HoYo.Test.Env
 import HoYo.Test.Utils
 
 import Control.Monad
@@ -8,5 +9,9 @@ import System.Exit
 
 main :: IO ()
 main = do
-  res <- and <$> sequence [bookmarkTests, utilsTests]
+  res <- and <$> sequence [
+                            bookmarkTests
+                          , utilsTests
+                          , envTests
+                          ]
   unless res exitFailure